cine_rag/requirements.txt
streamlit>=1.33.0
faiss-cpu>=1.7.4
chromadb>=0.4.0
sentence-transformers>=2.7.0
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
pytest>=8.0.0
